refactor : Names-> Players 이름 변경 및 플레이어 수 필드 추가
